+++ lib/src/commands/generate_feature_command.dart
@@ -111,10 +111,11 @@ class GenerateFeatureCommand extends Command<int> {
         return ExitCode.success.code;
       }
 
-      final confirm = _logger.confirm(
-        'Proceed with these changes?',
-        defaultValue: true,
-      );
+      final confirm = force ||
+          _logger.confirm(
+            'Proceed with these changes?',
+            defaultValue: true,
+          );
 
       if (!confirm) {
         _logger.info('Cancelled.');